refactor: 이미지 api 이동
